fix: modify resource plan os, cpu core, memory, disk size type, from int64 to float64. (merge request !1535)
Squash merge branch 'fix_float' into 'bcc/feature-plan'
fix: modify resource plan os, cpu core, memory, disk size type, from int64 to float64.

--story=119698450

 



TAPD: --story=119698450

+++ cmd/woa-server/logics/plan/ticket.go
@@ -94,8 +94,8 @@ func (c *Controller) CreateResPlanTicket(kt *kit.Kit, req *CreateResPlanTicketRe
 
 // constructResPlanTicket construct resource plan ticket.
 func constructResPlanTicket(req *CreateResPlanTicketReq, applicant string) (*rpt.ResPlanTicketTable, error) {
-	var originalOs, originalCpuCore, originalMemory, originalDiskSize int64
-	var updatedOs, updatedCpuCore, updatedMemory, updatedDiskSize int64
+	var originalOs, originalCpuCore, originalMemory, originalDiskSize float64
+	var updatedOs, updatedCpuCore, updatedMemory, updatedDiskSize float64
 	for _, demand := range req.Demands {
 		if demand.Original != nil {
 			originalOs += (*demand.Original).Cvm.Os
